renamed create_cat.py to create_example_cat.py and updated docustring
